+++ packages/partitura_core/lib/src/smufl/smufl_codepoints.dart
@@ -74,6 +74,8 @@ const Map<String, String> smuflCodepoints = {
   'timeSig7': '\uE087',
   'timeSig8': '\uE088',
   'timeSig9': '\uE089',
+  'timeSigCommon': '\uE08A',
+  'timeSigCutCommon': '\uE08B',
   'tuplet0': '\uE880',
   'tuplet1': '\uE881',
   'tuplet2': '\uE882',